Добавлен класс BaseActivity и в него вынесен код для изменения темы.
